Remove AdvancedUserInterface
