+++ src/Sylius/Component/User/Model/CredentialsHolderInterface.php
@@ -29,5 +29,5 @@ interface CredentialsHolderInterface extends PasswordAuthenticatedUserInterface
      * This is important if, at any given point, sensitive information like
      * the plain-text password is stored on this object.
      */
-    public function eraseCredentials();
+    public function eraseCredentials(): void;
 }
